+++ workspace/solutions/rust/src/s0/s3/s0/s3/mod.rs
@@ -1,0 +1,15 @@
+pub mod s_0303_00;
+pub mod s_0303_01;
+pub mod s_0303_test;
+
+pub trait ISolution {
+    fn init(nums: Vec<i32>) -> impl ISolution
+    where
+        Self: Sized;
+
+    fn solve(
+        &self,
+        left: i32,
+        right: i32,
+    ) -> i32;
+}